Update 2.3.2!
Improved screen box - Now appear also on self update
Fixed ip to players connecting from the same network as server but from other device - Thanks to @SinisterEmber
Added German translation - Thanks to @dadoirie (Dado)
Fixed self update which could only check for updates once
Added player nick to server console when player downloads modpack from AutoModpack Host
Added `generate_modpack_on_launch` option to config - thanks to which you will be able to disable the auto generation of modpack at the start of the server
Little icon change
Fixed errors when some config option were added
Fixed some malicious errors

+++ src/main/java/pl/skidam/automodpack/utils/SetupFiles.java
@@ -8,8 +8,8 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Objects;
 
-import static pl.skidam.automodpack.AutoModpackMain.selfOut;
 import static pl.skidam.automodpack.AutoModpackServer.changelogsDir;
+import static pl.skidam.automodpack.utils.JarUtilities.selfOut;
 
 public class SetupFiles {
     public SetupFiles() {
@@ -138,12 +138,14 @@ public class SetupFiles {
         }
 
         // extract icon
-        try {
-            new UnZipper(new File("./mods/" + selfOut.getName()), new File("./AutoModpack/"), "assets/automodpack/icon.png");
-            FileUtils.copyFileToDirectory(new File("./AutoModpack/assets/automodpack/icon.png"), new File("./AutoModpack/"));
-            FileUtils.deleteQuietly(new File("./AutoModpack/assets/"));
-        } catch (IOException e) {
-            e.printStackTrace();
+        if (!FabricLoader.getInstance().isDevelopmentEnvironment()) {
+            try {
+                new UnZipper(selfOut, new File("./AutoModpack/"), "assets/automodpack/icon.png");
+                FileUtils.copyFileToDirectory(new File("./AutoModpack/assets/automodpack/icon.png"), new File("./AutoModpack/"));
+                FileUtils.deleteQuietly(new File("./AutoModpack/assets/"));
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
     }
 }
